fix thread problem

+++ mindspore/lite/src/runtime/kernel/arm/fp32/addn_fp32.cc
@@ -44,7 +44,7 @@ int AddNCPUKernel::ReSize() { return RET_OK; }
 
 int AddNCPUKernel::AddNParallelRun(int thread_id) {
   int count_per_thread = UP_DIV(elements_num_, op_parameter_->thread_num_);
-  int count = MSMIN(count_per_thread, static_cast<int>(elements_num_ - thread_id * count_per_thread));
+  int count = MSMIN(count_per_thread, elements_num_ - thread_id * count_per_thread);
   auto stride = count_per_thread * thread_id;
   auto ret = ElementAdd(in1_addr_ + stride, in2_addr_ + stride, out_addr_ + stride, count);
   if (ret != NNACL_OK) {
